errors
存在一些报错，先上传

+++ SampleCode-V5/android-sdk-v5-sample/src/main/java/dji/sampleV5/aircraft/network/models/ReceivedWaypoint.kt
@@ -9,6 +9,6 @@ data class ReceivedWaypoint(
     @SerializedName("lat") val lat: Double,      // 纬度
     @SerializedName("lng") val lng: Double,      // 经度
     @SerializedName("alt") val alt: Double,      // 相对高度 (米)
-    @SerializedName("speed") val speed: Float = 5.0f, // 飞行速度 (米/秒)，默认 5.0
+    @SerializedName("speed") val speed: Double = 5.0, // 飞行速度 (米/秒)，默认 5.0
     @SerializedName("id") val missionId: Long = 0L // 任务唯一ID，用于去重，建议服务器提供时间戳
 )
